fix(arborist): fix usage of path of custom registry
Properly handle the resolution of a path via a custom registry with a path element in the URL, when the path element is a prefix of a package name.

When a custom registry with a path element is used (e.g. http://localhost:3080/npm/), then the resolution of download URLs for packages starting with the path element (in this case `npm`, so e.g. `npm-run-path`) will fail as the path element gets omitted.

## References
Fixes #8736